Changed baseURL to HTTPS
I changed all occurances of http://api.discogs.com to https.
In line 186 of ClientTest.php is a test that checks the length of an image. The length had increased due to my changes. I changed this test to pass, but actually I have no idea if the new length is the correct one or not.
$this->assertSame(37966, strlen($response['image']));